refactor(gh): extract gh/gists submodule
Fourth of the 5-PR series splitting src/gh/mod.rs into per-resource
submodules (issue #301). Moves gist listing, node-id mapping, and
file-content fetch into gh/gists.rs: gist_list_plan,
gist_starred_list_plan, current_user_plan, gist_view_plan,
parse_gist_list_json, gist_node_id_map, merge_gist_node_id_maps,
fetch_gist_list_json(_with), fetch_gist_starred_list_json(_with),
fetch_current_user_login(_with), parse_starred_gist_ids,
fetch_gist_file_content(_with), along with their unit tests.

parse_gh_gists, GhGist/GhGistForkOf/GhGistFile, and raw_url_fetch_plan
stay in gh/mod.rs as shared kernel (also used by comments/forks and
the not-yet-extracted revisions cluster respectively); gists.rs reaches
them via `use super::{parse_gh_gists, raw_url_fetch_plan};`, matching
the convention established by gh/stars.rs, gh/comments.rs, and
gh/forks.rs. gh/mod.rs re-exports gists' public items by explicit
name, so none of the existing crate::gh::X call sites need to change.

No signature or behavior change.

Part of #301
